Improve dynamic for readme and license in setuptools plugin
Setuptools has more options for handling `readme` or `license` than the
ones offered by PEP 621, therefore it makes sense to handle them via
`dynamic`.

Here `readme` is implemented as a "expanded" `file:` directive that also
accepts the `content-type` key.

On the other hand, `license` is more complicated to deal with (and is likely to
change with PEP 639), so for the time being we just preserve the same
structure as setuptools already handles.