Place license-file in tool.setuptools instead of dynamic
Now that we clarified that PEP 621 does not specify anything about the
(non-standard) core metadata field `License-file` and that it is not
necessary to add `license` to `dynamic` to be able to automatically fill
it, there is no motivation for keeping this field in
`tool.setuptools.dynamic` instead of `tool.setuptools` (directly).

As such there is also little motivation to allow `license` in `dynamic`...
There is nothing else setuptools can do outside of what the standard already
allows.

Note that this is unlikely to change with PEP 639 (the draft currently
establishes that the backend can backfill `license-files` without
needing an explicit `dynamic` if it is not provided).